requirements.txt
plotly
matplotlib
numpy
geopandas
folium
streamlit_folium